refactor(groups): extract GroupLifecycleOps from GroupConversationService
R1 step 8: post-discussion lifecycle ops (follow-up/continue/close/read/
delete/list, pending approvals, ephemeral cleanup, terminal-state failure
handling, dynamic-agent tracking propagation) into one collaborator.
Constructed per-call (mirrors GroupAttachmentBinder's attachmentBinder())
since deploymentStore is field-injected and not yet populated when the
facade's own constructor runs; operationsInProgress and activeTokens stay
on the facade, shared by reference.

GroupConversationService: 1,885 -> 1,380 lines. 8 of R1's 10 steps done.
